break out serialize code
