+++ vsc-extension/src-shared/source-info.ts
@@ -80,63 +80,3 @@ export type SelectedFunctionTestCases = Record<string, MaybeSelectedTestCase>;
 
 /** Map of filename -> functionName -> selected test index */
 export type SelectedFileTestCases = Record<string, SelectedFunctionTestCases>;
-
-export function makeSerializable(
-  sources: SourceFileMap
-): SerializableSourceFileMap {
-  const printer = ts.createPrinter({ removeComments: true });
-  const resultList = Object.entries(sources).map(
-    ([key, sourceFileInfo]): [string, SerializableSourceFileInfo] => [
-      key,
-      makeSerializableSourceFile(sourceFileInfo, printer),
-    ]
-  );
-  return Object.fromEntries(resultList);
-}
-
-function makeSerializableSourceFile(
-  sourceFileInfo: SourceFileInfo,
-  printer: ts.Printer
-): SerializableSourceFileInfo {
-  const { fileName } = sourceFileInfo.sourceFile;
-
-  return {
-    sourceFile: {
-      fileName,
-    },
-    functions: sourceFileInfo.functions.map((fn) =>
-      makeSerialiableFunction(fn, sourceFileInfo, printer)
-    ),
-  };
-}
-
-function makeSerialiableFunction(
-  fn: ts.FunctionDeclaration,
-  sourceFileInfo: SourceFileInfo,
-  printer: ts.Printer
-): SerializableFunctionDeclaration {
-  const { name } = fn;
-  return {
-    name: name?.text,
-    declaration: printer.printNode(
-      ts.EmitHint.Unspecified,
-      fn,
-      sourceFileInfo.sourceFile
-    ),
-    parameters: fn.parameters.map((param) => {
-      return {
-        name:
-          param.name.kind === ts.SyntaxKind.Identifier
-            ? param.name.escapedText.toString()
-            : "<unknown>",
-        tempType: param.type
-          ? (printer.printNode(
-              ts.EmitHint.Unspecified,
-              param.type,
-              sourceFileInfo.sourceFile
-            ) as any)
-          : "object",
-      };
-    }),
-  };
-}
